requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.4.2
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
dnspython==1.16.0
filelock==3.0.12
Flask==1.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.2
PyJWT==2.0.1
pylint==2.6.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
toml==0.10.2
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1